Updated VisualPACT.py with additional options and better image overlay code

+++ VisualPACT/VisualPACT.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 import matplotlib
 matplotlib.use('Agg')
 import numpy as np
@@ -48,6 +49,20 @@ def make_video(image_folder,video_name):
     cv2.destroyAllWindows()
     video.release()
 
+def make_transparent_overlay(overlay):
+    h, w, c = overlay.shape
+    # append Alpha channel -- required for BGRA (Blue, Green, Red, Alpha)
+    overlay_transparent_bg = np.concatenate([overlay, np.full((h, w, 1), 255, dtype=np.uint8)], axis=-1)
+    # create a mask where white pixels ([255, 255, 255]) are True
+    threshold=128
+    white = np.all(overlay >= [threshold, threshold, threshold], axis=-1)
+    black=np.all(overlay <= [threshold, threshold, threshold], axis=-1)
+    # change the values of Alpha to 0 for all the white pixels
+    overlay_transparent_bg[white, -1] = 0
+    overlay_transparent_bg[black,:] = 0
+    overlay_transparent_bg[black,-1] = 255
+    return overlay_transparent_bg
+
 #USER INPUTS
 parser = argparse.ArgumentParser()
 parser.add_argument('transient_data_file',action='store')
@@ -58,6 +73,9 @@ parser.add_argument('--fps',dest='fps',action='store',type=int,default=5)
 parser.add_argument('--layer',dest='layer',action='store',type=int,default=0)
 parser.add_argument('--dpi',dest='dpi',action='store',type=int,default=100)
 parser.add_argument('--K',dest='use_kelvin',action='store',type=bool,default=False)
+parser.add_argument("--steady",dest='steady_state',action='store',type=bool,default=False)
+parser.add_argument('--font_scale',dest='font_scale',action='store',type=float,default=1)
+
 #READ PARSER ARGUUMENTS
 parser_args = parser.parse_args()
 transient_data_file = parser_args.transient_data_file
@@ -71,14 +89,19 @@ if fps < 1:
     sys.exit(2)
 layer = parser_args.layer
 dpi = parser_args.dpi
+font_scale = parser_args.font_scale
 kelvin_offset=273.15
 if(parser_args.use_kelvin):
     kelvin_offset=0
+steady_state=parser_args.steady_state
 #SET OUTPUT PATHS
 output_path = transient_data_file.rstrip(''.join(Path(transient_data_file).suffixes))
 output_name = os.path.basename(output_path)
 video_file = output_path+'.avi'
-frame_folder = output_path+'_frames/'
+if steady_state:
+    frame_folder = os.path.dirname(transient_data_file)+'/steadyframes/'
+else:
+    frame_folder = output_path+'_frames/'
 grid_rows, grid_cols = getDimensions(transient_data_file,layer)
 if(grid_rows==0 or grid_cols==0):
     print("ERROR: Invalid data file or nonexistant layer.")
@@ -89,9 +112,13 @@ print("video path: "+os.path.abspath(video_file))
 print("image folder path: "+os.path.abspath(frame_folder)+'/')
 #Delete any previous images folder and make a new one
 folder_path = Path(frame_folder)
-if folder_path.exists() and folder_path.is_dir():
-    shutil.rmtree(folder_path)
-os.makedirs(frame_folder)
+if not steady_state:
+    if folder_path.exists() and folder_path.is_dir():
+        shutil.rmtree(folder_path)
+    os.makedirs(frame_folder)
+else:
+    if (not folder_path.exists()):
+        os.makedirs(frame_folder)
 #Read data from file
 print("Reading file...",end="\r")
 df_l = readFormatInput(transient_data_file,grid_rows,grid_cols)
@@ -122,8 +149,10 @@ cm_subsection = np.linspace(start, stop, number_of_lines)
 colors = [ matplotlib.cm.jet(x) for x in cm_subsection ]  
 #Generate video Frames
 i=0
+sns.set(font_scale=font_scale)
 if(use_overlay):
     overlay =  cv2.imread(overlay_image,cv2.IMREAD_UNCHANGED)
+    overlay_transparent_bg = make_transparent_overlay(overlay)
 for index, row in df_l.iterrows():
     print("frame: "+str(i),end="\r")
     newRow = np.array(row)
@@ -131,10 +160,11 @@ for index, row in df_l.iterrows():
     plot = sns.heatmap(newRow,cmap=colors,xticklabels=False, yticklabels=False,cbar_kws={'label':'Temperature($^\circ C$)'}, vmin=tmin, vmax=tmax)
     plot.set_aspect("equal")
     if(use_overlay):
-        plot.imshow(overlay, alpha=0.3, zorder=1, aspect=plot.get_aspect(), interpolation='none', extent=plot.get_xlim()+plot.get_ylim())
+        plot.imshow(overlay_transparent_bg,  zorder=1, aspect=plot.get_aspect(), interpolation='none', extent=plot.get_xlim()+plot.get_ylim())
     plot.get_figure().savefig(f"{frame_folder}{output_name}_{i}.png",dpi=dpi,bbox_inches = 'tight',pad_inches = 0)
     plot.get_figure().clf()
     i+=1
 
-make_video(frame_folder,video_file)
+if not steady_state:
+    make_video(frame_folder,video_file)
 print("Done.                  ")
